ref(KUI-1290): extracted api ErrorUtils
